+++ packages/main/src/RPA/Robocloud/Secrets.py
@@ -4,8 +4,10 @@ import collections
 import copy
 import json
 import logging
+import os
 import traceback
 from abc import abstractmethod, ABCMeta
+from typing import Tuple
 
 import requests
 from cryptography.exceptions import InvalidTag
@@ -13,6 +15,8 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import padding, rsa
 from cryptography.hazmat.primitives.ciphers.aead import AESGCM
+
+
 from robot.libraries.BuiltIn import BuiltIn, RobotNotRunningError
 
 from RPA.core.helpers import import_by_name, required_env
@@ -49,8 +53,14 @@ class Secret(collections.abc.Mapping):
     def description(self):
         return self._desc
 
+    def update(self, kvpairs):
+        self._dict.update(kvpairs)
+
     def __getitem__(self, key):
         return self._dict[key]
+
+    def __setitem__(self, key, value):
+        self._dict[key] = value
 
     def __contains__(self, key):
         return key in self._dict
@@ -75,6 +85,10 @@ class BaseSecretManager(metaclass=ABCMeta):
     @abstractmethod
     def get_secret(self, secret_name):
         """Return ``Secret`` object with given name."""
+
+    @abstractmethod
+    def set_secret(self, secret: Secret):
+        """Set a secret with a new value."""
 
 
 class FileSecrets(BaseSecretManager):
@@ -119,6 +133,16 @@ class FileSecrets(BaseSecretManager):
             self.logger.error("Failed to load secrets file: %s", err)
             return {}
 
+    def save(self):
+        """Save the secrets JSON to disk."""
+        try:
+            with open(self.path, "w") as f:
+                if not isinstance(self.data, dict):
+                    raise ValueError("Invalid content format")
+                json.dump(self.data, f, indent=4)
+        except (IOError, ValueError) as err:
+            self.logger.error("Failed to save secrets file: %s", err)
+
     def get_secret(self, secret_name):
         """Get secret defined with given name from file.
 
@@ -131,6 +155,16 @@ class FileSecrets(BaseSecretManager):
             raise KeyError(f"Undefined secret: {secret_name}")
 
         return Secret(secret_name, "", values)
+
+    def set_secret(self, secret: Secret) -> None:
+        """Set the secret value in the local Vault
+        with the given ``Secret`` object.
+
+        :param secret:                 A ``Secret`` object.
+        :raises IOError, ValueError:   Writing the local vault failed.
+        """
+        self.data[secret.name] = dict(secret)
+        self.save()
 
 
 class RobocloudVault(BaseSecretManager):
@@ -192,10 +226,20 @@ class RobocloudVault(BaseSecretManager):
 
         return self.__public_bytes
 
-    def create_url(self, name):
+    def create_secret_url(self, name):
         """Create a URL for a specific secret."""
         return url_join(
             self._host, "secrets-v1", "workspaces", self._workspace, "secrets", name
+        )
+
+    def create_public_key_url(self):
+        return url_join(
+            self._host,
+            "secrets-v1",
+            "workspaces",
+            self._workspace,
+            "secrets",
+            "publicKey",
         )
 
     def get_secret(self, secret_name):
@@ -205,7 +249,7 @@ class RobocloudVault(BaseSecretManager):
         :returns:                       Secret object
         :raises RobocloudVaultError:    Error with API request or response payload
         """
-        url = self.create_url(secret_name)
+        url = self.create_secret_url(secret_name)
 
         try:
             response = requests.get(url, headers=self.headers, params=self.params)
@@ -254,6 +298,104 @@ class RobocloudVault(BaseSecretManager):
 
         return payload
 
+    def set_secret(self, secret: Secret) -> None:
+        """Set the secret value in the Vault. Note that the secret possibly
+        consists of multiple key-value pairs, which will all be overwritten
+        with the values given here. So don't try to update only one item
+        of the secret, update all of them.
+
+        :param secret: A ``Secret`` object
+        """
+        value, aes_iv, aes_key, aes_tag = self._encrypt_secret_value_with_aes(secret)
+        pub_key = self.get_publickey()
+        aes_enc = self._encrypt_aes_key_with_public_rsa(aes_key, pub_key)
+
+        payload = {
+            "description": secret.description,
+            "encryption": {
+                "authTag": aes_tag.decode(),
+                "encryptedAES": aes_enc.decode(),
+                "encryptionScheme": self.ENCRYPTION_SCHEME,
+                "iv": aes_iv.decode(),
+            },
+            "name": secret.name,
+            "value": value.decode(),
+        }
+
+        url = self.create_secret_url(secret.name)
+        try:
+            response = requests.put(url, headers=self.headers, json=payload)
+            response.raise_for_status()
+        except Exception as e:
+            self.logger.debug(traceback.format_exc())
+            if response.status_code == 403:
+                raise RobocloudVaultError(
+                    "Failed to set secret value. Does your token have write access?"
+                ) from e
+            raise RobocloudVaultError("Failed to set secret value.") from e
+
+    def get_publickey(self) -> bytes:
+        """Get the public key for AES encryption with the existing token."""
+        url = self.create_public_key_url()
+        try:
+            response = requests.get(url, headers=self.headers)
+            response.raise_for_status()
+        except Exception as e:
+            self.logger.debug(traceback.format_exc())
+            raise RobocloudVaultError(
+                "Failed to fetch public key. Is your token valid?"
+            ) from e
+
+        return response.content
+
+    @staticmethod
+    def _encrypt_secret_value_with_aes(
+        secret: Secret,
+    ) -> Tuple[bytes, bytes, bytes, bytes]:
+        def generate_aes_key() -> Tuple[bytes, bytes]:
+            aes_key = AESGCM.generate_key(bit_length=256)
+            aes_iv = os.urandom(16)
+            return aes_key, aes_iv
+
+        def split_auth_tag_from_encrypted_value(
+            encrypted_value: bytes,
+        ) -> Tuple[bytes, bytes]:
+            """AES auth tag is the last 16 bytes of the AES encrypted value.
+            Split the tag from the value, Cloud needs that.
+            """
+            aes_tag = encrypted_value[-16:]
+            trimmed_encrypted_value = encrypted_value[:-16]
+            return trimmed_encrypted_value, aes_tag
+
+        value = json.dumps(dict(secret)).encode()
+
+        aes_key, aes_iv = generate_aes_key()
+        encrypted_value = AESGCM(aes_key).encrypt(aes_iv, value, b"")
+        encrypted_value, aes_tag = split_auth_tag_from_encrypted_value(encrypted_value)
+
+        return (
+            base64.b64encode(encrypted_value),
+            base64.b64encode(aes_iv),
+            aes_key,
+            base64.b64encode(aes_tag),
+        )
+
+    @staticmethod
+    def _encrypt_aes_key_with_public_rsa(aes_key: bytes, public_rsa: bytes) -> bytes:
+        pub_decoded = base64.b64decode(public_rsa)
+        public_key = serialization.load_der_public_key(pub_decoded)
+
+        aes_enc = public_key.encrypt(
+            aes_key,
+            padding.OAEP(
+                mgf=padding.MGF1(algorithm=hashes.SHA256()),
+                algorithm=hashes.SHA256(),
+                label=None,
+            ),
+        )
+
+        return base64.b64encode(aes_enc)
+
 
 class Secrets:
     """`Secrets` is a library for interfacing secrets set in the Robocloud Vault
@@ -289,12 +431,21 @@ class Secrets:
     .. code-block:: robotframework
 
         *** Settings ***
+        Library    Collections
         Library    RPA.Robocloud.Secrets
 
         *** Tasks ***
         Reading secrets
-            ${secrets}=   Get Secret  swaglabs
-            Log Many      ${secrets}
+            ${secret}=    Get Secret  swaglabs
+            Log Many      ${secret}
+
+        Modifying secrets
+            ${secret}=          Get Secret      swaglabs
+            ${level}=           Set Log Level   NONE
+            Set To Dictionary   ${secret}       username    nobody
+            Set Log Level       ${level}
+            Set Secret          ${secret}
+
 
     **Python**
 
@@ -302,9 +453,21 @@ class Secrets:
 
         from RPA.Robocloud.Secrets import Secrets
 
-        secret = Secrets()
-        print(f"My secrets: {secret.get_secret('swaglabs')}")
+        SECRETS = Secrets()
+
+
+        def reading_secrets():
+            print(f"My secrets: {SECRETS.get_secret('swaglabs')}")
+
+
+        def modifying_secrets():
+            secret = SECRETS.get_secret("swaglabs")
+            secret["username"] = "nobody"
+            secrets.set_secret(secret)
+
     """
+
+    # TODO: set to dictionary WILL log the values to log.html!
 
     ROBOT_LIBRARY_SCOPE = "GLOBAL"
     ROBOT_LIBRARY_DOC_FORMAT = "REST"
@@ -352,10 +515,20 @@ class Secrets:
 
         return factory
 
-    def get_secret(self, secret_name):
+    def get_secret(self, secret_name: str) -> Secret:
         """Read a secret from the configured source, e.g. Robocloud Vault,
         and return it as a ``Secret`` object.
 
         :param secret_name: Name of secret
         """
         return self.adapter.get_secret(secret_name)
+
+    def set_secret(self, secret: Secret) -> None:
+        """Overwrite an existing secret with new values.
+
+        Note: Only allows modifying existing secrets, and replaces
+              all values contained within it.
+
+        :param secret: Secret as a ``Secret`` object, from e.g. ``Get Secret``
+        """
+        self.adapter.set_secret(secret)