+++ docs/source/releasenotes.rst
@@ -4,6 +4,13 @@ Release notes
 Upcoming release
 ----------------
 
+8.2.0
+-----
+
+- Library **RPA.Robocloud.Secrets**:
+
+  - Add keyword ``Set Secret`` for updating stored secrets
+
 8.1.0
 -----